This is a program to implement a magic directory with build dictionary and search methods.
For the method buildDictionary, you'll be given a list of non-repetitive
words to build a dictionary.

For the method search, you'll be given a word, and judge whether if you
modify exactly one character into another character in this word, the
modified word is in the dictionary you just built.