This is a program to find the longest uncommon subsequence of this group of two strings.
The longest uncommon subsequence is defined as the longest subsequence
of one of these strings and this subsequence should not be any
subsequence of the other strings.

A subsequence is a sequence that can be derived from one sequence by
deleting some characters without changing the order of the remaining
elements. Trivially, any string is a subsequence of itself and an empty
string is a subsequence of any string.